+++ main.py
@@ -22,7 +22,7 @@ from PySide2.QtWidgets import QApplication, QMainWindow, QFileDialog, QMessageBo
 from PySide2.QtCore import QSize, QTranslator, QLibraryInfo, Qt
 
 from ui_main import Ui_MainWindow
-from db_handler import connectToDatabase, createElement, getElementIdByMetadata, insertInfo, editInfo, getElementInfo, getLastElementID, listofIDs, getElementMetadatabyID, erase_element, getImagesInfo, kill_connection, getLastImgs
+from db_handler import connectToDatabase, createElement, getElementIdByMetadata, insertInfo, editInfo, getElementInfo, getLastElementID, listofIDs, getElementMetadatabyID, erase_element, getImagesInfo, kill_connection, getLastImgs, getDocumentTypeByID
 from camcontrol import Cam
 import configparser
 import ctypes
@@ -50,6 +50,7 @@ else:
     sys.exit(0)
 
 # Entorno de la aplicación
+
 
 def restart():
     QtCore.QCoreApplication.quit()
@@ -109,11 +110,8 @@ class MainWindow(QMainWindow):
         # Página de metadatos
         widgets.tipodocComboBox.currentIndexChanged.connect(self.indexChange)
         widgets.enviarFormButton.clicked.connect(self.enviarForm)
-        widgets.enviarFormEditButton.clicked.connect(
-            self.editarForm)
         widgets.browserDirButton.clicked.connect(self.getDirName)
         widgets.zoom_dial.valueChanged.connect(lambda: self.zoom_dial())
-
 
         # abrir directorio
         widgets.openFolderButton.clicked.connect(self.openimagesDir)
@@ -138,7 +136,8 @@ class MainWindow(QMainWindow):
             if len(dir) != 0:
                 for files in dir:
                     if files != 'old_backup':
-                        shutil.move(os.path.join(IMGDIR, files), os.path.join(IMGDIR, 'old_backup', files))
+                        shutil.move(os.path.join(IMGDIR, files),
+                                    os.path.join(IMGDIR, 'old_backup', files))
                 log.log(f"Se hizo un backup de las capturas")
 
     # navigation functions
@@ -156,7 +155,8 @@ class MainWindow(QMainWindow):
             try:
                 Cams.pause_devs()
             except Exception as e:
-                log.log(f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+                log.log(
+                    f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
             widgets.stackedWidget.setCurrentWidget(widgets.inicioPage)
             self.display_elements()
         elif btnName == "backtoInicioButton":
@@ -164,7 +164,8 @@ class MainWindow(QMainWindow):
             try:
                 Cams.pause_devs()
             except Exception as e:
-                log.log(f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+                log.log(
+                    f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
             msgBox = QMessageBox()
             msgBox.setIcon(QMessageBox.Warning)
             msgBox.setText("¿Está seguro que desea salir del formulario?")
@@ -181,7 +182,8 @@ class MainWindow(QMainWindow):
             try:
                 Cams.pause_devs()
             except Exception as e:
-                log.log(f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+                log.log(
+                    f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
             widgets.stackedWidget.setCurrentWidget(widgets.metadataPage)
             widgets.tipoColeccion.setCurrentWidget(widgets.formLegajo)
             widgets.botones_metadata.setCurrentWidget(widgets.enviar)
@@ -220,7 +222,7 @@ class MainWindow(QMainWindow):
             element_name = elemento[1]
             element_description = elemento[2]
 
-            image_path = Path(IMGDIR, f'{element_id}','data', 'JPG')
+            image_path = Path(IMGDIR, f'{element_id}', 'data', 'JPG')
             image_not_found = "imgs/No-Photo-Available.png"
 
             # Display image in the grid
@@ -265,12 +267,13 @@ class MainWindow(QMainWindow):
                           QSize(), QIcon.Normal, QIcon.Off)
             button.setIcon(icon1)
             button.setIconSize(QSize(20, 20))
-            button.clicked.connect(self.edit_element)
-            widgets.elementslayout.addWidget(button, id, 2)
+            button.clicked[bool].connect(
+                lambda _, element_id=element_id: self.edit_element(element_id))
+            widgets.elementslayout.addWidget(button, id, 3)
 
             button = QPushButton()
             # button.setText("Añadir imágenes")
-            button.setObjectName(f"edit_element_{element_id}")
+            button.setObjectName(f"add_more_{element_id}")
             button.setMinimumSize(QSize(42, 42))
             button.setMaximumSize(QSize(42, 42))
             icon1 = QIcon()
@@ -278,7 +281,8 @@ class MainWindow(QMainWindow):
                           QSize(), QIcon.Normal, QIcon.Off)
             button.setIcon(icon1)
             button.setIconSize(QSize(20, 20))
-            button.clicked[bool].connect(lambda _, element_id=element_id: self.set_scanner_page(element_id))
+            button.clicked[bool].connect(
+                lambda _, element_id=element_id: self.set_scanner_page(element_id))
             widgets.elementslayout.addWidget(button, id, 2)
 
             button = QPushButton()
@@ -291,8 +295,9 @@ class MainWindow(QMainWindow):
                           QSize(), QIcon.Normal, QIcon.Off)
             button.setIcon(icon2)
             button.setIconSize(QSize(20, 20))
-            button.clicked[bool].connect(lambda _, element_id=element_id, image_path=image_path: self.export_element(element_id))
-            widgets.elementslayout.addWidget(button, id, 3)
+            button.clicked[bool].connect(
+                lambda _, element_id=element_id, image_path=image_path: self.export_element(element_id))
+            widgets.elementslayout.addWidget(button, id, 4)
 
             button = QPushButton()
             # button.setText("Eliminar")
@@ -304,8 +309,9 @@ class MainWindow(QMainWindow):
                           QSize(), QIcon.Normal, QIcon.Off)
             button.setIcon(icon3)
             button.setIconSize(QSize(20, 20))
-            button.clicked[bool].connect(lambda _, element_id=element_id, image_path=image_path: self.delete_element(element_id, image_path))
-            widgets.elementslayout.addWidget(button, id, 4)
+            button.clicked[bool].connect(
+                lambda _, element_id=element_id, image_path=image_path: self.delete_element(element_id, image_path))
+            widgets.elementslayout.addWidget(button, id, 5)
 
         widgets.verticalLayout_20.addLayout(widgets.elementslayout)
 
@@ -315,7 +321,9 @@ class MainWindow(QMainWindow):
         '''
         # go to metadata page
         widgets.stackedWidget.setCurrentWidget(widgets.metadataPage)
-        widgets.botones_metadata.setCurrentWidget(widgets.editar)
+        dtype = getDocumentTypeByID(element_id)
+
+        widgets.tipodocComboBox.setCurrentIndex(dtype)
 
         # write element_id in the form
         self.write_fields_info(element_id)
@@ -554,7 +562,7 @@ class MainWindow(QMainWindow):
         and write in its fields the information retrieved from the db
         '''
 
-        tipo_documento = getElementInfo(element_id)['tipo_documento']
+        tipo_documento = getElementInfo(element_id)['document_type']
         data = getElementMetadatabyID(element_id)
 
         if tipo_documento == 1:
@@ -618,6 +626,11 @@ class MainWindow(QMainWindow):
         elif tipo_documento == 6:
             QMessageBox.information(
                 self, 'Error', 'No se puede editar un escaneo sencillo')
+
+        # assesst editar button
+        # get current widget
+        widgets.botones_metadata.setCurrentWidget(widgets.editar)
+        widgets.enviarFormEditButton.clicked[bool].connect(lambda _, element_id=element_id: self.editarForm(element_id))
 
     def cleanForm(self, tipo_de_documento):
         '''
@@ -693,7 +706,7 @@ class MainWindow(QMainWindow):
         tipo_de_documento = widgets.tipodocComboBox.currentIndex() + 1
 
         #zoom_value = widgets.zoom_dial.value()
-        zoom_value =  widgets.zoom_valuedit.text()
+        zoom_value = widgets.zoom_valuedit.text()
 
         # dng_checkbox value
         if widgets.dng_check.isChecked():
@@ -808,7 +821,8 @@ class MainWindow(QMainWindow):
             return "dual_camera_mode"
 
         except IndexError as e:
-            log.log(f"ERROR: {e} en {__file__} linea {e.__traceback__.tb_lineno}")
+            log.log(
+                f"ERROR: {e} en {__file__} linea {e.__traceback__.tb_lineno}")
             # single camera QMessageBox, accept retry cancel
             respuesta = QMessageBox().question(self, "Una cámara conectada",
                                                "Solamente una cámara está conectada.\n \
@@ -850,7 +864,8 @@ class MainWindow(QMainWindow):
             os.makedirs(folder_path, exist_ok=True)
         except OSError as e:
             print(e)
-            log.log(f'ERROR: Al crear {folder_path} se encontró un OSError {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+            log.log(
+                f'ERROR: Al crear {folder_path} se encontró un OSError {e} en {__file__} linea {e.__traceback__.tb_lineno}')
             raise
 
         # set folder_path to label
@@ -870,10 +885,11 @@ class MainWindow(QMainWindow):
         # get last image number pair
         last_img_group = getLastImgs(element_id)
         try:
-            numbers = [li.replace('.jpg', '').replace('.dng', '') for li in last_img_group]
+            numbers = [li.replace('.jpg', '').replace('.dng', '')
+                       for li in last_img_group]
             numbers = list(dict.fromkeys(numbers))
             # esto evita que se presenten errores si se desconecta una cámara y se sigue capturando
-            # con una de ellas. 
+            # con una de ellas.
             # TODO: Con una sola cámara ¿cómo evitar que se creen ids para pares o impares?
             numbers = numbers[:2]
             last_img_left = [int(i) for i in numbers if int(i) % 2 != 0]
@@ -890,24 +906,27 @@ class MainWindow(QMainWindow):
                     last_img_left = {last_img_left}')
             elif len(last_img_right) == 0:
                 try:
-                    last_img_right = last_img_left -1
+                    last_img_right = last_img_left - 1
                 except TypeError:
                     last_img_right = last_img_left[0] - 1
                 log.log(f'WARNING: Se ha restablecido la secuencia en la cámara derecha. \
                     last_img_right = {last_img_right}')
             else:
                 # cachar algún error desconocido
-                log.log(f'INFO: secuencia de imágenes sin issues. last_img_left = {last_img_left} y last_img_right = {last_img_right}')
+                log.log(
+                    f'INFO: secuencia de imágenes sin issues. last_img_left = {last_img_left} y last_img_right = {last_img_right}')
                 pass
             try:
-                last_img_left = [last_img_left[0] + 2 if last_img_left is not 0 else 1][0]
+                last_img_left = [last_img_left[0] +
+                                 2 if last_img_left is not 0 else 1][0]
             except TypeError:
-                last_img_left = [last_img_left + 2 if last_img_left is not 0 else 1][0]
+                last_img_left = [last_img_left +
+                                 2 if last_img_left is not 0 else 1][0]
             last_img_left = f'{last_img_left:04d}'
             try:
                 last_img_right = last_img_right[0] + 2
             except TypeError:
-                last_img_right = last_img_right +2
+                last_img_right = last_img_right + 2
             last_img_right = f'{last_img_right:04d}'
         except AttributeError:
             last_img_left = '0001'
@@ -919,7 +938,6 @@ class MainWindow(QMainWindow):
             widgets.directorio_elementos.text(), 'data', 'JPG', f'{last_img_left}.jpg')
         right_img_path = Path(
             widgets.directorio_elementos.text(), 'data', 'JPG', f'{last_img_right}.jpg')
-        
 
         try:
             zoom = int(config['camaras']['zoom_predeterminado'])
@@ -929,11 +947,11 @@ class MainWindow(QMainWindow):
         try:
             # TODO: ¡Hacer que esto funcione!
             dng_status = config['DEFAULT']['dng']
-            Cams.captura(element_ident, last_img_left, last_img_right, zoom, dng_status)
+            Cams.captura(element_ident, last_img_left,
+                         last_img_right, zoom, dng_status)
         except TypeError:
             QMessageBox().warning(self, "Error",
                                         "No se encontraron cámaras", QMessageBox.Ok)
-
 
         # Este loop permite
         if Cams.len_devs() > 1:
@@ -946,7 +964,7 @@ class MainWindow(QMainWindow):
         else:
             time.sleep(5)
 
-        # for some reason QPixmap is not working with PosixPath 
+        # for some reason QPixmap is not working with PosixPath
         # raise TypeError. As a temporary solution transform PosixPath to str
         left_img_path = left_img_path.absolute().as_posix()
         right_img_path = right_img_path.absolute().as_posix()
@@ -958,7 +976,8 @@ class MainWindow(QMainWindow):
 
         widgets.statusLabel.setText(
             f"Capturadas {last_img_left} y {last_img_right}")
-        log.log(f'INFO: Imágenes {last_img_left} y {last_img_right} capturadas')
+        log.log(
+            f'INFO: Imágenes {last_img_left} y {last_img_right} capturadas')
 
         # display validation buttons
         widgets.controlesCamstackedWidget.setCurrentWidget(
@@ -974,7 +993,7 @@ class MainWindow(QMainWindow):
     def resetCaptura(self):
         # delete most recent images from directory
         folder_path = widgets.directorio_elementos.text()
-        
+
         tipos = ['JPG', 'DNG']
         for t in tipos:
             last_imagenes = glob.glob(f'{folder_path}/{t}/*.{t.lower()}')
@@ -984,7 +1003,8 @@ class MainWindow(QMainWindow):
                 try:
                     os.remove(f)
                 except OSError as e:
-                    log.log(f'ERROR: Al eliminar {f} se encontró un OSError {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+                    log.log(
+                        f'ERROR: Al eliminar {f} se encontró un OSError {e} en {__file__} linea {e.__traceback__.tb_lineno}')
                     raise
 
         # erase image labels
@@ -1013,7 +1033,8 @@ class MainWindow(QMainWindow):
             try:
                 Cams.pause_devs()
             except Exception as e:
-                log.log(f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
+                log.log(
+                    f'WARNING: No se pudieron detener las cámaras. {e} en {__file__} linea {e.__traceback__.tb_lineno}')
 
     def gentle_close(self):
         '''